controller: Move controller/grpc/protobuf -> controller/protobuf
