--- controller/grpc/.gitignore
@@ -1,1 +1,0 @@
-/controller-grpc
